Merging r226754: ------------------------------------------------------------------------ r226754 | hans | 2015-01-21 16:45:20 -0800 (Wed, 21 Jan 2015) | 9 lines
Fix compiler_builtins.m test to not rely on including system stdlib.h and malloc.h

Importing _Builtin_intrinsics.sse and avx would transitively pull in those
headers, and the test would fail when building in an environment where
they were not available on the include path.

This fixes PR20995 for me.

Differential Revision: http://reviews.llvm.org/D7112
------------------------------------------------------------------------


git-svn-id: http://llvm.org/svn/llvm-project/cfe/branches/release_36@226833 91177308-0d34-0410-b5e6-96231b3b80d8